create validation module and validate gossip blocks
