+++ crates/astria-sequencer-relayer/Cargo.toml
@@ -28,7 +28,7 @@ tracing = { workspace = true }
 tracing-subscriber = { workspace = true, features = ["env-filter"] }
 tokio = { workspace = true, features = [ "macros", "rt-multi-thread" ] }
 
-astria-gossipnet = { path = "../astria-gossipnet", features = [ "mdns" ] }
+astria-gossipnet = { path = "../astria-gossipnet" }
 astria-proto = { path = "../astria-proto" }
 astria-rs-cnc = { path = "../astria-rs-cnc" }
